--- src/main/java/TaskParseException.java
@@ -1,5 +1,0 @@
-public class TaskParseException extends Exception {
-    public TaskParseException(String message) {
-        super(message);
-    }
-}
