+++ src/main/java/duke/Storage.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -6,12 +8,27 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
 
+import duke.exceptions.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
+/**
+ * Saves and loads tasks to hard drive in String form
+ */
 class Storage {
+    /**
+     * name of file and path which the data is to be stored
+     */
     static final String FILE_NAME = "./data/duke.txt";
 
+    /**
+     * check if file that is to be written to and loaded from exists, if
+     * it does not, create one
+     */
     static void checkFileExists() {
         Path path = Paths.get(FILE_NAME);
         try {
@@ -26,31 +43,35 @@ class Storage {
         }
     }
 
-    public static void saveTask(List<Task> tasks) {
+    /**
+     * wipes file and save tasks given in param to hard drive
+     * @param tasks ArrayList of tasks to be saved
+     */
+    public static void saveTask(ArrayList<Task> tasks) {
         checkFileExists();
         try (FileWriter fileWriter = new FileWriter(FILE_NAME)) {
             for (Task task : tasks) {
                 String text = "";
                 if (task instanceof Todo) {
-                    String done = task.isDone
+                    String done = task.getisDone()
                             ? "1"
                             : "0";
-                    String desc = task.description;
+                    String desc = task.getDescription();
                     text = "T|" + done + "|" + desc;
                 } else if (task instanceof Event) {
-                    String done = task.isDone
+                    String done = task.getisDone()
                             ? "1"
                             : "0";
-                    String desc = task.description;
-                    LocalDate from = ((Event) task).from;
-                    LocalDate to = ((Event) task).to;
+                    String desc = task.getDescription();
+                    LocalDate from = ((Event) task).getFrom();
+                    LocalDate to = ((Event) task).getTo();
                     text = "E|" + done + "|" + desc + "|" + from + "|" + to;
                 } else if (task instanceof Deadline) {
-                    String done = task.isDone
+                    String done = task.getisDone()
                             ? "1"
                             : "0";
-                    String desc = task.description;
-                    LocalDate by = ((Deadline) task).by;
+                    String desc = task.getDescription();
+                    LocalDate by = ((Deadline) task).getBy();
                     text = "D|" + done + "|" + desc + "|" + by;
                 }
                 fileWriter.write(text + "\n");
@@ -60,7 +81,11 @@ class Storage {
         }
     }
 
-
+    /**
+     * loads task from hard drive and returns it
+     * @return ArrayList of tasks
+     * @throws DukeException if there are no tasks in the hard drive
+     */
     public static ArrayList<Task> loadTasks() throws DukeException {
         checkFileExists();
         ArrayList<Task> tasks = new ArrayList<>();
