+++ src/main/java/duke/exceptions/TaskParseException.java
@@ -1,0 +1,14 @@
+package duke.exceptions;
+
+/**
+ * Represents a parsing error between String and Task
+ */
+public class TaskParseException extends Exception {
+    /**
+     * public onstructor, returns a <code>TaskParseException</code> instance
+     * @param message message that is to be displayed
+     */
+    public TaskParseException(String message) {
+        super(message);
+    }
+}
